FossilOrigin-Name: e20fcb5159f2a74ea60ac6a0287165d114fc71d7510a6d9068962f239dde619f

+++ test/kvtest.c
@@ -562,7 +562,7 @@ static int exportMain(int argc, char **argv){
     nWrote = fwrite(pData, 1, (size_t)nData, out);
     fclose(out);
     printf("\r%s   ", zTail); fflush(stdout);
-    if( nWrote!=nData ){
+    if( nWrote!=(size_t)nData ){
       fatalError("Wrote only %d of %d bytes to %s\n",
                   (int)nWrote, nData, zFN);
     }
